feat(dashboard): switch / over to the snapshot dashboard
Operator-approved changeover after the /dashboard-preview review: the real
dashboard now IS the decision-first snapshot page, and the temporary preview
route is gone. No page other than the dashboard is touched.

* routes/+page.svelte renders entirely from GET /api/dashboard/snapshot —
  KPI strip on the truth rules, needs-attention inbox, money → machine →
  research tiers, browser-side last-good retention with explicit
  OFFLINE/STALE/FETCH FAILED/UNAVAILABLE states. The legacy panel fan-out
  (3 aggregate loader calls + 14 self-fetching panels, incl. the duplicate
  alert surfaces) is gone from the page; panel components stay on disk per
  the locked plan (phase-3 cleanup decides their fate).
* routes/+page.ts keeps only ssr=false and the legacy ?view= 301 redirects;
  the three-endpoint loader is removed.
* /dashboard-preview route deleted; lib/components/dashboard_preview/ →
  lib/components/dashboard_snapshot/ (naming now matches the backend module
  and store); testids preview-* → dash-*.
* dashboardPreview.test.ts folds into dashboardRoute.test.ts: page render,
  wrapper truth states, inbox, store retention, and the preserved legacy
  ?view= redirect tests replace the panel-era route test.

Verified at this state: svelte-check 0 errors; frontend vitest 437/437;
ruff (CI scope) clean; full backend suite in repo .venv 6,295 passed /
9 skipped / 1 failed (test_duplicate_registration_guard::
test_different_params_allowed — passes 6/6 solo, order-dependent flake,
backend diff of this commit is empty).
