+++ frontend/src/lib/components/dashboard_snapshot/AgentsTile.svelte
