+++ frontend/src/lib/components/dashboard_snapshot/AttentionInbox.svelte
