+++ frontend/src/lib/components/dashboard_snapshot/DataTile.svelte
