+++ frontend/src/lib/components/dashboard_snapshot/EquityTile.svelte
