+++ frontend/src/lib/components/dashboard_snapshot/KpisStrip.svelte
