+++ frontend/src/lib/components/dashboard_snapshot/LeaderboardTile.svelte
