+++ frontend/src/lib/components/dashboard_snapshot/PaperTile.svelte
