+++ frontend/src/lib/components/dashboard_snapshot/PipelineTile.svelte
