+++ frontend/src/lib/components/dashboard_snapshot/SchedulerTile.svelte
