+++ frontend/src/lib/components/dashboard_snapshot/SnapshotSection.svelte
