+++ frontend/src/lib/components/dashboard_snapshot/SystemTile.svelte
