+++ frontend/src/lib/components/dashboard_snapshot/TradingTile.svelte
