+++ frontend/src/lib/components/dashboard_snapshot/format.ts
