--- frontend/src/routes/dashboard-preview/+page.ts
@@ -1,1 +1,0 @@
-export const ssr = false;
